+++ src/app/components/inputs/select-dropdown-input/select-dropdown-input.component.ts
@@ -1,7 +1,7 @@
 import {ApiErrors} from '../../../models/api-models/api-errors/api-errors';
 import {Component} from '@angular/core';
 import {EventEmitter} from '@angular/core';
-import {getNestedProperty} from '../../../utils/object-util';
+import {getNestedProperty} from '../../../utils/object/object.util';
 import {Input} from '@angular/core';
 import {InputErrorsComponent} from '../../form-errors/input-errors/input-errors.component';
 import {Output} from '@angular/core';
