+++ src/app/components/inputs/skills-input/skills-input.component.ts
@@ -1,6 +1,6 @@
 import {ApiErrors} from '../../../models/api-models/api-errors/api-errors';
 import {Component} from '@angular/core';
-import {deleteElementFromArray} from '../../../utils/array-util';
+import {deleteElementFromArray} from '../../../utils/array/array.util';
 import {FormControl} from '@angular/forms';
 import {Input} from '@angular/core';
 import {Skill} from '../../../models/api-models/skill/skill';
