+++ src/app/components/job-list-item/job-list-item.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {Job} from '../../models/api-models/job/job';
 import {NavigationService} from '../../services/navigation.service';
-import {JARoutes} from '../../routes/ja-routes';
+import {JARoutes} from '../../routes/ja-routes/ja-routes';
 
 @Component({
   selector: 'job-list-item',
