+++ src/app/components/job-map-marker/job-map-marker.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {Job} from '../../models/api-models/job/job';
 import {NavigationService} from '../../services/navigation.service';
-import {JARoutes} from '../../routes/ja-routes';
+import {JARoutes} from '../../routes/ja-routes/ja-routes';
 
 @Component({
   selector: 'job-map-marker',
