+++ src/app/components/menus/navigation-menu/navigation-menu.component.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {EventEmitter} from '@angular/core';
 import {Input} from '@angular/core';
-import {JARoutes} from '../../../routes/ja-routes';
+import {JARoutes} from '../../../routes/ja-routes/ja-routes';
 import {NavigationService} from '../../../services/navigation.service';
 import {NavigationStart} from '@angular/router';
 import {OnDestroy} from '@angular/core';
