+++ src/app/components/navbars/app-navbar/app-navbar.component.ts
@@ -2,7 +2,7 @@ import {Component} from '@angular/core';
 import {Input} from '@angular/core';
 import {EventEmitter} from '@angular/core';
 import {Output} from '@angular/core';
-import {JARoutes} from '../../../routes/ja-routes';
+import {JARoutes} from '../../../routes/ja-routes/ja-routes';
 import {Language} from '../../../models/api-models/language/language';
 import {SystemLanguagesResolver} from '../../../resolvers/system-languages/system-languages.resolver';
