+++ src/app/guards/not-logged-in/not-logged-in.guard.ts
@@ -1,7 +1,7 @@
 import {ActivatedRouteSnapshot} from '@angular/router';
 import {CanActivate} from '@angular/router';
 import {Injectable} from '@angular/core';
-import {JARoutes} from '../../routes/ja-routes';
+import {JARoutes} from '../../routes/ja-routes/ja-routes';
 import {NavigationService} from '../../services/navigation.service';
 import {RouterStateSnapshot} from '@angular/router';
 import {UserResolver} from '../../resolvers/user/user.resolver';
