+++ src/app/resolvers/user/user.resolver.ts
@@ -1,7 +1,7 @@
 import {DataStoreService} from '../../services/data-store.service';
 import {EventEmitter} from '@angular/core';
 import {Injectable} from '@angular/core';
-import {JARoutes} from '../../routes/ja-routes';
+import {JARoutes} from '../../routes/ja-routes/ja-routes';
 import {NavigationService} from '../../services/navigation.service';
 import {Resolve} from '@angular/router';
 import {User} from '../../models/api-models/user/user';
