+++ src/app/routes/ja-route/ja-route.ts
