+++ src/app/services/api-call.service.ts
@@ -5,10 +5,10 @@ import {environment} from '../../environments/environment';
 import {Headers} from '@angular/http';
 import {Http} from '@angular/http';
 import {Injectable} from '@angular/core';
-import {JARoutes} from '../routes/ja-routes';
+import {JARoutes} from '../routes/ja-routes/ja-routes';
 import {NavigationService} from './navigation.service';
 import {Observable} from 'rxjs';
-import {parseJsonapiResponse} from '../utils/jsonapi-parser.util';
+import {parseJsonapiResponse} from '../utils/jsonapi-parser/jsonapi-parser.util';
 import {Request} from '@angular/http';
 import {RequestMethod} from '@angular/http';
 import {RequestOptions} from '@angular/http';
