+++ src/app/services/navigation.service.ts
@@ -1,7 +1,7 @@
 import {ActivatedRoute} from '@angular/router';
 import {Injectable} from '@angular/core';
-import {JARoute} from '../routes/ja-route';
-import {JARoutes} from '../routes/ja-routes';
+import {JARoute} from '../routes/ja-route/ja-route';
+import {JARoutes} from '../routes/ja-routes/ja-routes';
 import {NavigationCancel} from '@angular/router';
 import {NavigationEnd} from '@angular/router';
 import {NavigationError} from '@angular/router';
