+++ src/app/storage/local-storage/local-storage.ts
@@ -1,4 +1,4 @@
-import {StorageInterface} from './storage-interface';
+import {StorageInterface} from '../storage-interface/storage-interface';
 
 export class LocalStorage implements StorageInterface {
   private store: any;
