+++ src/app/storage/memory-storage/memory-storage.ts
@@ -1,4 +1,4 @@
-import {StorageInterface} from './storage-interface';
+import {StorageInterface} from '../storage-interface/storage-interface';
 
 export class MemoryStorage implements StorageInterface {
   private items: any = {};
