+++ src/app/utils/ascii-mapper/ascii-mapper.util.ts
