+++ src/app/utils/data-url/data-url.util.ts
