+++ src/app/utils/storage-type-available/storage-type-available.util.ts
