--- src/app/utils/string-util.ts
@@ -1,5 +1,0 @@
-export function compareStrings(a: string, b: string): number {
-  if (a < b) return -1;
-  if (a > b) return 1;
-  return 0;
-}
