+++ src/app/utils/string/string.util.ts
@@ -1,4 +1,10 @@
-import {AsciiMapper} from './ascii-mapper';
+import {AsciiMapper} from '../ascii-mapper/ascii-mapper.util';
+
+export function compareStrings(a: string, b: string): number {
+  if (a < b) return -1;
+  if (a > b) return 1;
+  return 0;
+}
 
 export function isValidSSNCharCode(charCode): boolean {
   const isNumber = charCode >= AsciiMapper.zero && charCode <= AsciiMapper.nine;
